+++ CHANGELOG.md
@@ -22,6 +22,9 @@ project adheres to [Semantic Versioning](http://semver.org/).
 - Fix bug where `browser.addArguments` JSON config setting did not work for
   Firefox.
 
+- Add `browser.profile` JSON config setting that sets the browser profile
+  directory. Currently supported in Chrome and Firefox.
+
 - Upgrade dependencies.
 
 ## [0.5.9] 2021-04-22